requirements.txt
appdirs==1.4.4
asgiref==3.4.1
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
billiard==3.6.4.0
black==19.3b0
boto==2.49.0
boto3==1.12.48
botocore==1.15.49
CacheControl==0.12.6
cachetools==4.2.2
celery==5.1.2
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.4.0
django-import-export==2.5.0
django-oauth-toolkit==1.3.2
django-redis==4.11.0
django-request==1.6.1
django-ses==0.8.14
django-sql-explorer==1.1.3
django-storages==1.9.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.1.0
Faker==3.0.0
future==0.18.2
gunicorn==19.10.0
httplib2==0.19.1
idna==2.10
importlib-metadata==1.7.0
ipython-genutils==0.2.0
jedi==0.17.1
jmespath==0.10.0
kombu==5.1.0
MarkupPy==1.14
more-itertools==8.4.0
msgpack==1.0.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.4
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
phonenumbers==8.12.1
pickleshare==0.7.5
pika==1.1.0
Pillow==7.1.2
pluggy==0.13.1
prompt-toolkit==3.0.5
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.2.1
pytest-black==0.3.7
pytest-django==3.6.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.4.1
redis==3.5.3
requests==2.23.0
rsa==4.7.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.0
tablib==3.0.0
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.9
vine==5.0.0
watchdog==0.10.3
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.1.0
black==19.3b0
django-debug-toolbar==2.2
ipython==7.15.0
pytest-testmon==0.9.19
pytest-watch==4.2.0
pytest-cov==2.8.1
twilio==7.6.0
